DEPLOYMENT_PACKAGE/requirements.txt
streamlit
pandas
psycopg2-binary
google-generativeai
python-dotenv
